requirements.txt
telethon
tweepy
python-dotenv
schedule
tweet-splitter
pillow